Data conversion improved and bug fixed in it
This fixes a bug which can lead to interger overflow if data is close to
extrema. This has lead to incorporate the offset value in addition to the
scaling factor which in turn allows to convert unsigned integer to/from
signed integer.
bugfix: integer overflow when converting a data close to minimal value
bugfix: Data conversion intermediate is sometime badly chosen

+++ tests/testbdf.c
@@ -10,6 +10,7 @@
 #include "filecmp.h"
 #include "copy_xdf.h"
 
+#define RAMP_NS		50
 #define SAMPLINGRATE	128
 #define DURATION	13
 #define NITERATION	((SAMPLINGRATE*DURATION)/NSAMPLE)
@@ -17,34 +18,69 @@
 #define NEEG	11
 #define NEXG	7
 #define NTRI	1
-#define scaled_t	float
-static const enum xdftype arrtype = XDFFLOAT;
-static const enum xdftype trigarrtype = XDFINT32;
+
+const char sess_str[] = "This a test BDF file";
+const char subj_str[] = "Nobody. This string is very long on purpose and test the truncation. It should be longer than the length of the field in the file";
+
+typedef float	eeg_t;
+#define EEG_TYPE 	XDFFLOAT
+typedef double	sens_t;
+#define SENS_TYPE 	XDFDOUBLE
+typedef uint32_t	tri1_t;
+#define TRI1_TYPE	XDFUINT32
+#define TRI1_MIN	0.0
+#define TRI1_MAX        16777216.0
+typedef int32_t		tri2_t;
+#define TRI2_TYPE	XDFINT32
+#define TRI2_MIN	-8388608.0
+#define TRI2_MAX        8388607.0 
 off_t offskip[2] = {168, 184};
 
+#define PMIN (-262144.0)
+#define PMAX (262143.0)
 
-void WriteSignalData(scaled_t* eegdata, scaled_t* exgdata, int32_t* tridata, int seed)
+void set_signal_values(eeg_t* eeg, sens_t* exg, tri1_t* tri1, tri2_t* tri2)
 {
-	int i,j;
+	int i,j, is, ir;
+	double dv;
 	static int isample = 0;
 
 	for(i=0; i<NSAMPLE; i++) {
-		for (j=0; j<NEEG; j++)	
-			eegdata[i*NEEG+j] = ((i+isample)%23)/*((j+1)*i)+seed*/;
-	}
-	for(i=0; i<NSAMPLE; i++) {
-		for (j=0; j<NEXG; j++)	
-			exgdata[i*NEXG+j] = ((j+1)*i)+seed;
-	}
-	for(i=0; i<NSAMPLE; i++) {
-		for (j=0; j<NTRI; j++)	
-			tridata[i*NTRI+j] = (((i+isample)%10 == 0)?6:0);
+		is = i+isample;
+		ir = is % RAMP_NS;
+
+		for (j=0; j<NEEG; j++)	{
+			dv = ir / (double)(RAMP_NS-1);
+			dv = dv*(PMAX-PMIN) + PMIN;
+			eeg[i*NEEG+j] = dv / (j+1);
+		}
+
+		for (j=0; j<NEXG; j++) {	
+			dv = ir / (double)(RAMP_NS-1);
+			dv = dv*(PMAX-PMIN) + PMIN;
+			exg[i*NEXG+j] = dv / (j+1);
+		}
+
+		for (j=0; j<NTRI; j++) {
+			tri1[i*NTRI+j] = 0;
+			if (ir == 0)
+				tri1[i*NTRI+j] = (is/RAMP_NS % 2) 
+						? 131072 : 4096;
+		}
+
+		for (j=0; j<NTRI; j++) {
+			tri2[i*NTRI+j] = 0;
+			if (ir == 0)
+				tri2[i*NTRI+j] = (is/RAMP_NS % 2) 
+						? -256 : 256 ;
+		}
 	}
 	isample += NSAMPLE;
 }
 
 
-static int set_default_analog(struct xdf* xdf, int arrindex)
+static int set_default_analog(struct xdf* xdf, int arrindex,
+						enum xdftype arrtype)
 {
 	xdf_set_conf(xdf, 
 		XDF_CF_ARRTYPE, arrtype,
@@ -61,16 +97,18 @@ static int set_default_analog(struct xdf* xdf, int arrindex)
 	return 0;
 }
 
-static int set_default_trigger(struct xdf* xdf, int arrindex)
+static int set_default_trigger(struct xdf* xdf, int arrindex,
+						enum xdftype arrtype,
+						double pmin, double pmax)
 {
 	xdf_set_conf(xdf, 
-		XDF_CF_ARRTYPE, trigarrtype,
+		XDF_CF_ARRTYPE, arrtype,
 		XDF_CF_ARRINDEX, arrindex,
 		XDF_CF_ARROFFSET, 0,
 		XDF_CF_TRANSDUCTER, "Triggers and Status",
 		XDF_CF_PREFILTERING, "No filtering",
-		XDF_CF_PMIN, -8388608.0,
-		XDF_CF_PMAX, 8388607.0,
+		XDF_CF_PMIN, pmin,
+		XDF_CF_PMAX, pmax,
 		XDF_CF_UNIT, "Boolean",
 		XDF_CF_RESERVED, "TRI",
 		XDF_NOF);
@@ -81,102 +119,113 @@ static int set_default_trigger(struct xdf* xdf, int arrindex)
 
 int generate_xdffile(const char* filename)
 {
-	scaled_t* eegdata;
-	scaled_t* exgdata;
-	int32_t* tridata;
-	int phase;
+	eeg_t* eegdata;
+	sens_t* exgdata;
+	tri1_t* tri1data;
+	tri2_t* tri2data;
+	int retcode = -1;
 	struct xdf* xdf = NULL;
 	int i,j;
 	char tmpstr[16];
-	unsigned int strides[3] = {
+	unsigned int strides[4] = {
 		NEEG*sizeof(*eegdata),
 		NEXG*sizeof(*exgdata),
-		NTRI*sizeof(*tridata)
+		NTRI*sizeof(*tri1data),
+		NTRI*sizeof(*tri2data),
 	};
 
-
-	phase = 5;
 
 	// Allocate the temporary buffers for samples
 	eegdata = malloc(NEEG*NSAMPLE*sizeof(*eegdata));
 	exgdata = malloc(NEXG*NSAMPLE*sizeof(*exgdata));
-	tridata = malloc(NTRI*NSAMPLE*sizeof(*tridata));
-	if (!eegdata || !exgdata || !tridata)
+	tri1data = malloc(NTRI*NSAMPLE*sizeof(*tri1data));
+	tri2data = malloc(NTRI*NSAMPLE*sizeof(*tri2data));
+	if (!eegdata || !exgdata || !tri1data || !tri2data)
 		goto exit;
 		
 
-	phase--;
 	xdf = xdf_open(filename, XDF_WRITE, XDF_BDF);
 	if (!xdf) 
 		goto exit;
+	xdf_set_conf(xdf, XDF_F_SAMPLING_FREQ, (int)SAMPLINGRATE,
+	                  XDF_F_SESS_DESC, sess_str,
+			  XDF_F_SUBJ_DESC, subj_str,
+			  XDF_NOF);
 	
 	// Specify the structure (channels and sampling rate)
-	phase--;
-	xdf_set_conf(xdf, XDF_F_SAMPLING_FREQ, (int)SAMPLINGRATE, XDF_NOF);
-	set_default_analog(xdf, 0);
+	set_default_analog(xdf, 0, EEG_TYPE);
 	for (j=0; j<NEEG; j++) {
 		sprintf(tmpstr, "EEG%i", j);
 		if (!xdf_add_channel(xdf, tmpstr))
 			goto exit;
 	}
 
-	xdf_set_conf(xdf, XDF_CF_ARRINDEX, 1, XDF_CF_ARROFFSET, 0, XDF_NOF);
+	set_default_analog(xdf, 1, SENS_TYPE);
 	for (j=0; j<NEXG; j++) {
 		sprintf(tmpstr, "EXG%i", j);
 		if (!xdf_add_channel(xdf, tmpstr))
 			goto exit;
 	}
 
-	set_default_trigger(xdf, 2);
+	set_default_trigger(xdf, 2, TRI1_TYPE, TRI1_MIN, TRI1_MAX);
 	for (j=0; j<NTRI; j++) {
-		sprintf(tmpstr, "TRI%i", j);
+		sprintf(tmpstr, "TRI1%i", j);
+		if (!xdf_add_channel(xdf, tmpstr))
+			goto exit;
+	}
+
+	set_default_trigger(xdf, 3, TRI2_TYPE, TRI2_MIN, TRI2_MAX);
+	for (j=0; j<NTRI; j++) {
+		sprintf(tmpstr, "TRI2%i", j);
 		if (!xdf_add_channel(xdf, tmpstr))
 			goto exit;
 	}
 
 
 	// Make the the file ready for accepting samples
-	phase--;	
-	xdf_define_arrays(xdf, 3, strides);
+	xdf_define_arrays(xdf, 4, strides);
 	if (xdf_prepare_transfer(xdf) < 0)
 		goto exit;
 	
 	// Feed with samples
-	phase--;
 	for (i=0; i<NITERATION; i++) {
 		// Set data signals and unscaled them
-		WriteSignalData(eegdata, exgdata, tridata, i);
-		if (xdf_write(xdf, NSAMPLE, eegdata, exgdata, tridata) < 0)
+		set_signal_values(eegdata, exgdata, tri1data, tri2data);
+		if (xdf_write(xdf, NSAMPLE, eegdata, exgdata, tri1data, tri2data) < 0)
 			goto exit;
 	}
-	phase--;
+	retcode = 0;
 
 exit:
 	// if phase is non zero, a problem occured
-	if (phase) 
+	if (retcode) 
 		fprintf(stderr, "\terror: %s\n", strerror(errno));
 
 	// Clean the structures and ressources
 	xdf_close(xdf);
 	free(eegdata);
 	free(exgdata);
-	free(tridata);
+	free(tri1data);
+	free(tri2data);
 
-	return phase;
+	return retcode;
 }
 
 
 
 int main(int argc, char *argv[])
 {
-	int retcode = 0, keep_file = 0, opt;
+	int retcode = 0, keep_file = 0, opt, testcopy = 1;
 	char genfilename[] = "essaiw.bdf";
 	char reffilename[128];
 
-	while ((opt = getopt(argc, argv, "k")) != -1) {
+	while ((opt = getopt(argc, argv, "kc:")) != -1) {
 		switch (opt) {
 		case 'k':
 			keep_file = 1;
+			break;
+		case 'c':
+			testcopy = atoi(optarg);
 			break;
 
 		default:	/* '?' */
@@ -192,26 +241,27 @@ int main(int argc, char *argv[])
 	// Create the filename for the reference
 	snprintf(reffilename, sizeof(reffilename),
 		 "%s/ref%u-%u-%u-%u-%u-%u-%u.bdf", getenv("srcdir"),SAMPLINGRATE, DURATION,
-		 NITERATION, NSAMPLE, NEEG, NEXG, NTRI);
+		 NITERATION, RAMP_NS, NEEG, NEXG, NTRI);
 
 	
 	// Test generation of a file
 	unlink(genfilename);
 	retcode = generate_xdffile(genfilename);
 	if (!retcode)
-		retcode = cmp_files(genfilename, reffilename, 1, offskip);
+		retcode = cmp_files(genfilename, reffilename, 1, offskip, NULL);
 
 	// Test copy a file (implied reading)
-	unlink(genfilename);
-	retcode = copy_xdf(genfilename, reffilename, XDF_BDF);
-	if (!retcode)
-		retcode = cmp_files(genfilename, reffilename, 1, offskip);
-
+	if (!retcode && testcopy) {
+		unlink(genfilename);
+		retcode = copy_xdf(genfilename, reffilename, XDF_BDF);
+		if (!retcode)
+			retcode = cmp_files(genfilename, reffilename, 1, offskip, NULL);
+	}
 
 	if (!keep_file)
 		unlink(genfilename);
 
 
-	return retcode;
+	return retcode ? EXIT_FAILURE : EXIT_SUCCESS;
 }
